requirements.txt
flask
flask-cors
requests
numpy
gunicorn
opencv-python-headless
tensorflow-cpu
pillow